+++ R/lazy_bootstrap.R
@@ -1,48 +1,4 @@
-# df = iris[1:4]
-# names(df) <- letters[1:4]
-# res_mx <- mx_profiles(data = df, classes = 1:4)
-#
-# srmr <- function(x, y){
-#   select_these <- lower.tri(x)
-#   sqrt(mean((y[select_these] - x[select_these])^2))
-# }
-#
-# vc_sim = cov(sims)
-# vc_obs = cov(thisres$data)
-# for(j in 1:ncol(vc_obs)){
-#   for(k in 1:(j-1)){
-#     print(((vc_sim[j,k]/sqrt(vc_sim[j,j]*vc_sim[k,k]))-(vc_obs[j,k]/sqrt(vc_obs[j,j]*vc_obs[k,k])))^2)
-#   }
-# }
-#
-# srmr.MxModel <- function(null, alt, statistics = "all"){
-#   if(!inherits(y, what = "MxModel")) stop("Both 'x' and 'y' arguments must be of the same type.")
-#
-# }
-#
-#
-# srmr.data.frame <- function(null, alt, statistics = "all"){
-#   cor_null <- cor(null)
-#   cor_alt <- cor(alt)
-#   select_these <- lower.tri(cor_null)
-#   sd_null <- apply(null, 2, sd)
-#   sd_alt <- apply(alt, 2, sd)
-#   vec <- vector(mode = "numeric")
-#   if(any(c("cov", "all") %in% statistics)){
-#     vec <- append(vec, (cor_alt[select_these] - cor_null[select_these]))
-#   }
-#   if(any(c("var", "all") %in% statistics)){
-#     vec <- append(vec, (sd_alt^2 - sd_null^2)/sd_alt^2)
-#   }
-#   if(any(c("means", "all") %in% statistics)){
-#     vec <- append(vec, (colMeans(alt)/sd_alt - colMeans(null)/sd_null))
-#   }
-#   sqrt(mean(vec)^2)
-# }
-#
-# mr <- function(x){
-#   sqrt(mean(x[lower.tri(x)]^2))
-# }
+globalVariables("y")
 
 #' @title Predictive Model Comparison
 #' @description Compare (non-nested) models estimated on the same data using
@@ -72,7 +28,7 @@ pmc_srmr.mixture_list <- function(x, ..., reps = 100, ci = .95){
   select_these <- lower.tri(cor_obs)
   cor_obs_sel <- cor_obs[select_these]
 
-  if(requireNamespace("progressr", quietly = TRUE) & requireNamespace("progress", quietly = TRUE)){
+  if(requireNamespace("future.apply", quietly = TRUE) & requireNamespace("progressr", quietly = TRUE) & requireNamespace("progress", quietly = TRUE)){
     progressr::with_progress({
       pgs <- progressr::progressor(steps = reps*length(x))
       rep_stat <- do.call(cbind, lapply(seq_along(x), function(i){
@@ -109,8 +65,8 @@ pmc_srmr.mixture_list <- function(x, ..., reps = 100, ci = .95){
 
   }
 
-  srsm_meds <- apply(rep_stat, 2, median)
-  names(srsm_meds) <- nams
+  srmr_meds <- apply(rep_stat, 2, median)
+  names(srmr_meds) <- nams
 
   cis <- list(
     dif_seq = rep_stat[, 2:dim(rep_stat)[2]] - rep_stat[, 1:(dim(rep_stat)-1L)[2]],
@@ -124,30 +80,118 @@ pmc_srmr.mixture_list <- function(x, ..., reps = 100, ci = .95){
                alt = rep(nams[2:length(nams)], 2))
   out <- data.frame(comparison = rep(c("dif_seq", "dif_one"), each = nrow(out)/2),
                     out)
-  out$null_srsm <- srsm_meds[out$null]
-  out$alt_srsm <- srsm_meds[out$alt]
+  out$null_srmr <- srmr_meds[out$null]
+  out$alt_srmr <- srmr_meds[out$alt]
   out <- data.frame(out, cis, sig = c("", "*")[(apply(sign(cis), 1, sum) == -2)+1L])
   return(out)
 }
 
-#' #' @importFrom nonnest2 llcont
-#' ic_confint <- function(x, y, conf.level = .95){
-#'   ll1 <- nonnest2::llcont(x)
-#'   ll2 <- nonnest2::llcont(y)
-#'   n <- x$data$numObs
-#'   omega.hat.2 <- (n - 1)/n * var(ll1 - ll2, na.rm = TRUE)
-#'   bic1 <- AIC(object1, k = log(n))
-#'   aic1 <- AIC(object1)
-#'   bic2 <- AIC(object2, k = log(n))
-#'   aic2 <- AIC(object2)
-#'   bicdiff <- bic1 - bic2
-#'   aicdiff <- aic1 - aic2
-#'   alpha <- 1 - conf.level
-#'   BICci <- bicdiff + qnorm(c(alpha/2, (1 - alpha/2))) * sqrt(n *
-#'                                                                4 * omega.hat.2)
-#'   AICci <- aicdiff + qnorm(c(alpha/2, (1 - alpha/2))) * sqrt(n *
-#'                                                                4 * omega.hat.2)
-#'   out <- c(bic1, bic2, BICci, aic1, aic2, AICci)
-#'   names(out) <- c(paste0("BIC", 1:2), paste0("dBIC_", c("lb", "ub")), paste0("AIC", 1:2), paste0("dAIC_", c("lb", "ub")))
-#'   return(out)
+# #' @importFrom nonnest2 llcont
+# ic_confint <- function(x, y, conf.level = .95){
+#   ll1 <- nonnest2::llcont(x)
+#   ll2 <- nonnest2::llcont(y)
+#   n <- x$data$numObs
+#   omega.hat.2 <- (n - 1)/n * var(ll1 - ll2, na.rm = TRUE)
+#   bic1 <- AIC(object1, k = log(n))
+#   aic1 <- AIC(object1)
+#   bic2 <- AIC(object2, k = log(n))
+#   aic2 <- AIC(object2)
+#   bicdiff <- bic1 - bic2
+#   aicdiff <- aic1 - aic2
+#   alpha <- 1 - conf.level
+#   BICci <- bicdiff + qnorm(c(alpha/2, (1 - alpha/2))) * sqrt(n *
+#                                                                4 * omega.hat.2)
+#   AICci <- aicdiff + qnorm(c(alpha/2, (1 - alpha/2))) * sqrt(n *
+#                                                                4 * omega.hat.2)
+#   out <- c(bic1, bic2, BICci, aic1, aic2, AICci)
+#   names(out) <- c(paste0("BIC", 1:2), paste0("dBIC_", c("lb", "ub")), paste0("AIC", 1:2), paste0("dAIC_", c("lb", "ub")))
+#   return(out)
+# }
+
+
+#' @title Calculate Standardized Root Mean Residual
+#' @description Given two datasets, computes the correlation matrix for both,
+#' and then calculates the standardized root mean residual difference between
+#' these two correlation matrices.
+#' @param x An object for which a method of [stats::cor()] exists
+#' (e.g, `data.frame`).
+#' @param y An object for which a method of [stats::cor()] exists
+#' (e.g, `data.frame`).
+#' @return `numeric`
+#' @examples
+#' \dontrun{
+#' if(interactive()){
+#'  srmr(iris[1:2], iris[3:4])
 #' }
+#' }
+#' @rdname srmr
+#' @export
+srmr <- function(x, y){
+  cor_null <- cor(x)
+  cor_alt <- cor(y)
+  select_these <- lower.tri(cor_null)
+  dif <- cor_alt[select_these] - cor_null[select_these]
+  sqrt(mean(dif)^2)
+}
+
+
+#' @rdname pmc_srmr
+#' @param FUN Function used to compare the real data (referred to as `x`) to the
+#' model-implied data (referred to as `y`).
+#' @export
+pmc <- function(x, ..., reps = 20, ci = .95, FUN = srmr(x, y)){
+  UseMethod("pmc", x)
+}
+
+#' @method pmc mixture_list
+#' @export
+pmc.mixture_list <- function(x, ..., reps = 100, ci = .95, FUN = srmr(x, y)){
+  FUN <- match.fun(FUN)
+  nams <- do.call(c, lapply(x, function(i) i@name))
+  eval_env <- new.env()
+  assign("x", x[[1]]$data$observed, envir = eval_env)
+
+  if(requireNamespace("future.apply", quietly = TRUE) & requireNamespace("progressr", quietly = TRUE) & requireNamespace("progress", quietly = TRUE)){
+    progressr::with_progress({
+      pgs <- progressr::progressor(steps = reps*length(x))
+      rep_stat <- do.call(cbind, lapply(seq_along(x), function(i){
+        progmsg <- x[[i]]@name
+        simres <- do.call(c, future.apply::future_replicate(
+          n = reps,
+          future.seed = TRUE,
+          simplify = FALSE,
+          expr = {
+            pgs(sprintf(progmsg))
+            assign("y", OpenMx::mxGenerateData(x[[i]]), envir = eval_env)
+            browser()
+            as.numeric(eval(FUN, envir = eval_env))
+          }
+        ))
+      }))
+    })
+  } else {
+    rep_stat <- simplify2array(lapply(seq_along(x), function(i) {
+      replicate(reps, {
+        assign("y", OpenMx::mxGenerateData(x[[i]]), envir = eval_env)
+        as.numeric(eval(body(FUN), envir = eval_env))
+      })}))
+  }
+  meds <- apply(rep_stat, 2, median)
+  names(meds) <- nams
+
+  cis <- list(
+    dif_seq = rep_stat[, 2:dim(rep_stat)[2]] - rep_stat[, 1:(dim(rep_stat)-1L)[2]],
+    dif_one = rep_stat[, 2:dim(rep_stat)[2]] - rep_stat[, rep(1, (dim(rep_stat)[2]-1L))]
+  )
+  cis <- do.call(rbind, lapply(cis, function(i){matrix(as.vector(apply(i, 2, stats::quantile, probs = c(((1-ci)/2), 1-((1-ci)/2)))), ncol = 2, byrow = TRUE)}))
+  colnames(cis) <- c("lb", "ub")
+  out <- data.frame(
+    null = c(nams[1:(length(nams)-1L)], rep(nams[1], (length(nams)-1L))),
+    alt = rep(nams[2:length(nams)], 2))
+  out <- data.frame(comparison = rep(c("dif_seq", "dif_one"), each = nrow(out)/2),
+                    out)
+  out$null <- meds[out$null]
+  out$alt <- meds[out$alt]
+  out <- data.frame(out, cis, sig = c("", "*")[(apply(sign(cis), 1, sum) == -2)+1L])
+  return(out)
+}